Fix/prt 1882 (#79)
* fix: improve code

* feat: report config style
